# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de carteiras digitais multi-tenant desenvolvido em Go",
        "title": "Suno Wallets API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ops/dedup/candidates": {
            "get": {
                "description": "Retorna lista paginada de candidatos de deduplicação detectados, com filtros opcionais por CPF e status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations Dedup"
                ],
                "summary": "Listar candidatos de deduplicação",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "Filtrar por CPF específico",
                        "name": "cpf",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "PENDING",
                        "description": "Filtrar por status (PENDING, RESOLVED, IGNORED)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de candidatos de deduplicação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ops/dedup/resolve": {
            "post": {
                "description": "Resolve candidatos de deduplicação aplicando ações específicas (manter, remover, mesclar). Requer autenticação de admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations Dedup"
                ],
                "summary": "Resolver candidatos de deduplicação (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chave secreta de admin (se configurada)",
                        "name": "X-Admin-Secret",
                        "in": "header"
                    },
                    {
                        "description": "Parâmetros de resolução",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resolução aplicada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos ou erro na resolução",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - requer permissão de admin",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ops/dedup/scan": {
            "post": {
                "description": "Executa varredura para detectar possíveis transações duplicadas no sistema. Requer autenticação de admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations Dedup"
                ],
                "summary": "Executar scan de deduplicação (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chave secreta de admin (se configurada)",
                        "name": "X-Admin-Secret",
                        "in": "header"
                    },
                    {
                        "description": "Parâmetros do scan de deduplicação",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado do scan - candidatos criados",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - requer permissão de admin",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro na execução do scan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "TenantID": {
            "description": "ID do inquilino para multi-tenancy",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        }
    }
}